+++ src/wazuh_testing/utils/manage_agents.py
@@ -11,7 +11,7 @@ from typing import List
 from wazuh_testing.constants.api import AGENTS_ROUTE
 from wazuh_testing.constants.paths.binaries import MANAGE_AGENTS_BINARY
 from wazuh_testing.modules.api.utils import login, get_base_url
-from wazuh_testing.utils.db_queries.global_db import delete_agent
+from wazuh_testing.utils.db_queries.global_db import delete_agent, list_agents_ids
 
 
 def remove_agents(agents_id: List, remove_type: str = 'wazuhdb') -> None:
@@ -31,7 +31,7 @@ def remove_agents(agents_id: List, remove_type: str = 'wazuhdb') -> None:
     if agents_id:
         for agent_id in agents_id:
             if remove_type == 'manage_agents':
-                subprocess.call([MANAGE_AGENTS_BINARY, "-r", agent_id], stdout=open(os.devnull, "w"),
+                subprocess.call([MANAGE_AGENTS_BINARY, "-r", f"{agent_id}"], stdout=open(os.devnull, "w"),
                                 stderr=subprocess.STDOUT)
             elif remove_type == 'wazuhdb':
                 delete_agent(agent_id)
@@ -47,3 +47,7 @@ def remove_agents(agents_id: List, remove_type: str = 'wazuhdb') -> None:
             response_data = response.json()
             if response.status_code != 200:
                 raise RuntimeError(f"Error deleting an agent: {response_data}")
+
+
+def remove_all_agents(remove_type):
+    remove_agents(list_agents_ids(), remove_type)
